refactor(app): move filesystem CRUD handlers into useFileSystemManager
Move handleCreateNode, handleRenameNode, handleDeleteNode,
handleMoveNode, handleCut, handleCopy, and handlePaste out of App.tsx
into useFileSystemManager.ts, alongside the tree-manipulation
primitives the hook already had. App.tsx drops to 4480 lines.

Part of the App.tsx maintainability refactor (bmf-vangard-renpy-ide-cfo).
