+++ src/App.tsx
@@ -55,6 +55,7 @@ import { useSettingsManagement } from '@/hooks/useSettingsManagement';
 import { useFileSystemState } from '@/hooks/useFileSystemState';
 import { useStickyNotes } from '@/hooks/useStickyNotes';
 import { useProjectIO, type PendingStoryLayoutRefresh, type PendingRouteLayoutRefresh } from '@/hooks/useProjectIO';
+import { useFileSystemManager } from '@/hooks/useFileSystemManager';
 import { formatErrorMessage } from '@/lib/formatErrorMessage';
 import {
   computeStoryLayout,
@@ -2912,132 +2913,13 @@ const App: React.FC = () => {
     }
   }, [addToast]);
 
-  const handleCreateNode = useCallback(async (parentPath: string, name: string, type: 'file' | 'folder') => {
-    if (!window.electronAPI || !projectRootPath) return;
-    try {
-        const fullPath = await window.electronAPI.path.join(projectRootPath, parentPath, name);
-        if (type === 'folder') {
-            await window.electronAPI.createDirectory(fullPath);
-        } else {
-            await window.electronAPI.writeFile(fullPath, '');
-
-            // If it's an .rpy file, create a corresponding block
-            if (name.toLowerCase().endsWith('.rpy')) {
-                const relativePath = parentPath ? `${parentPath}/${name}` : name;
-                const content = ''; // Empty content for newly created files
-                addBlock(relativePath, content);
-                addToast(`Created block for ${name}`, 'success');
-            }
-        }
-        const projData = await window.electronAPI.loadProject(projectRootPath);
-        setFileSystemTree(projData.tree);
-    } catch (err) {
-        logger.error('Failed to create file/folder:', err);
-        addToast(`Failed to create ${type}: ${name}`, 'error');
-    }
-  }, [projectRootPath, addBlock, addToast, setFileSystemTree]);
-
-  const handleRenameNode = useCallback(async (oldPath: string, newName: string) => {
-      if (!window.electronAPI || !projectRootPath) return;
-      try {
-          const fullOldPath = await window.electronAPI.path.join(projectRootPath, oldPath) as string;
-          const parentDir = oldPath.split('/').slice(0, -1).join('/');
-          const fullNewPath = await window.electronAPI.path.join(projectRootPath, parentDir, newName) as string;
-          await window.electronAPI.moveFile(fullOldPath, fullNewPath);
-          const projData = await window.electronAPI.loadProject(projectRootPath);
-          setFileSystemTree(projData.tree);
-      } catch (err) {
-          logger.error('Failed to rename:', err);
-          addToast('Failed to rename file', 'error');
-      }
-  }, [projectRootPath, addToast, setFileSystemTree]);
-
-  const handleDeleteNode = useCallback(async (paths: string[]) => {
-      if (!window.electronAPI || !projectRootPath) return;
-      
-      // Check if any of the paths are .rpy files that have corresponding blocks
-      const rpyFilesToDelete = paths.filter(path => path.toLowerCase().endsWith('.rpy'));
-      const blocksToDelete = rpyFilesToDelete.map(rpyPath => 
-          blocks.find(block => block.filePath === rpyPath)
-      ).filter(Boolean) as Block[];
-      
-      // Show confirmation modal
-      openDeleteConfirmModal(paths, async () => {
-              try {
-                  // Delete the files
-                  for (const p of paths) {
-                      const fullPath = await window.electronAPI.path.join(projectRootPath, p) as string;
-                      await window.electronAPI.removeEntry(fullPath);
-                  }
-
-                  // Remove corresponding blocks for .rpy files
-                  blocksToDelete.forEach(block => {
-                      if (block) {
-                          deleteBlock(block.id);
-                          addToast(`Removed block for ${block.filePath}`, 'info');
-                      }
-                  });
-
-                  const projData = await window.electronAPI.loadProject(projectRootPath);
-                  setFileSystemTree(projData.tree);
-
-                  if (blocksToDelete.length > 0) {
-                      addToast(`Deleted ${paths.length} file(s) and removed ${blocksToDelete.length} block(s)`, 'success');
-                  } else {
-                      addToast(`Deleted ${paths.length} file(s)`, 'success');
-                  }
-              } catch (err) {
-                  logger.error('Failed to delete:', err);
-                  addToast('Failed to delete file(s)', 'error');
-              }
-      });
-  }, [projectRootPath, blocks, deleteBlock, addToast, openDeleteConfirmModal, setFileSystemTree]);
-
-  const handleMoveNode = useCallback(async (sourcePaths: string[], targetPath: string) => {
-      if (!window.electronAPI || !projectRootPath) return;
-      try {
-          const fullTargetDir = await window.electronAPI.path.join(projectRootPath, targetPath);
-          for (const p of sourcePaths) {
-              const fullSource = await window.electronAPI.path.join(projectRootPath, p);
-              const fileName = p.split('/').pop() || '';
-              const fullDest = await window.electronAPI.path.join(fullTargetDir, fileName);
-              await window.electronAPI.moveFile(fullSource, fullDest);
-          }
-          const projData = await window.electronAPI.loadProject(projectRootPath);
-          setFileSystemTree(projData.tree);
-      } catch (err) {
-          logger.error('Failed to move file(s):', err);
-          addToast('Failed to move file(s)', 'error');
-      }
-  }, [projectRootPath, addToast, setFileSystemTree]);
-
-  const handleCut = useCallback((paths: string[]) => setClipboard({ type: 'cut', paths: new Set(paths) }), [setClipboard]);
-  const handleCopy = useCallback((paths: string[]) => setClipboard({ type: 'copy', paths: new Set(paths) }), [setClipboard]);
-  const handlePaste = useCallback(async (targetPath: string) => {
-      if (!clipboard || !window.electronAPI || !projectRootPath) return;
-      try {
-          const fullTargetDir = await window.electronAPI.path.join(projectRootPath!, targetPath);
-
-          for (const p of clipboard.paths) {
-              const fullSource = await window.electronAPI.path.join(projectRootPath!, p);
-              const fileName = p.split('/').pop() || '';
-              const fullDest = await window.electronAPI.path.join(fullTargetDir, fileName);
-
-              if (clipboard.type === 'cut') {
-                  await window.electronAPI.moveFile(fullSource, fullDest);
-              } else {
-                  await window.electronAPI.copyEntry(fullSource, fullDest);
-              }
-          }
-
-          if (clipboard.type === 'cut') setClipboard(null);
-          const projData = await window.electronAPI.loadProject(projectRootPath);
-          setFileSystemTree(projData.tree);
-      } catch (err) {
-          logger.error('Failed to paste:', err);
-          addToast('Failed to paste file(s)', 'error');
-      }
-  }, [clipboard, projectRootPath, addToast, setClipboard, setFileSystemTree]);
+  const {
+      handleCreateNode, handleRenameNode, handleDeleteNode, handleMoveNode,
+      handleCut, handleCopy, handlePaste,
+  } = useFileSystemManager({
+      projectRootPath, setFileSystemTree, blocks, addBlock, deleteBlock,
+      clipboard, setClipboard, openDeleteConfirmModal, addToast,
+  });
 
   // --- User Snippet CRUD ---
   const handleSaveSnippet = (snippet: UserSnippet) => {
